Finish migrating to typescript
