+++ client/package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "@types/jest": "^29.5.6",
     "@types/node": "^20.8.9",
-    "@types/react": "^18.2.32",
+    "@types/react": "^18.2.33",
     "@types/react-dom": "^18.2.14",
     "font-awesome": "^4.7.0",
     "gh-pages": "^2.0.1",
