+++ client/src/routes/root.tsx
@@ -1,4 +1,6 @@
-import Login from './../components/Login';
+import React from 'react';
+import Login from '../components/Login';
+import './../App.css';
 
 export default function Root() {
     return (
